+++ modules/mod_sparcc.R
@@ -1158,7 +1158,7 @@ mod_sparcc_server <- function(id, ps_obj) {
     })
 
     output$download_comparison_network_plot <- downloadHandler(
-      filename = function() paste0("sparcc_differential_network_", Sys.Date(), ".png"),
+      filename = function() paste0("sparcc_differential_network_", format(Sys.time(), "%Y%m%d_%H%M%S"), ".png"),
       content = function(file) {
         req(network_result())
         validate(need(identical(input$analysis_mode, "Compare two groups"), "Enable 'Compare two groups' to download differential network."))
@@ -1431,7 +1431,7 @@ mod_sparcc_server <- function(id, ps_obj) {
     })
 
     output$download_network_plot_all <- downloadHandler(
-      filename = function() paste0("sparcc_network_all_", Sys.Date(), ".png"),
+      filename = function() paste0("sparcc_network_all_", format(Sys.time(), "%Y%m%d_%H%M%S"), ".png"),
       content = function(file) {
         req(network_result())
         plot_width <- as.integer(input$plot_width)
@@ -1446,7 +1446,7 @@ mod_sparcc_server <- function(id, ps_obj) {
     )
 
     output$download_network_plot_connected <- downloadHandler(
-      filename = function() paste0("sparcc_network_connected_", Sys.Date(), ".png"),
+      filename = function() paste0("sparcc_network_connected_", format(Sys.time(), "%Y%m%d_%H%M%S"), ".png"),
       content = function(file) {
         req(network_result())
         plot_width <- as.integer(input$plot_width)